+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod fs;
 pub mod hints;
 pub mod inode;
 pub mod sparse_cache;
+pub mod spool;